bio updated
